@@ -28,7 +28,7 @@ export default defineComponent({
     <div class="count">
       <pre class="hug"><code>count</code></pre>
       <p class="hug">(original)</p>
-      <button @click="count--">+</button> {{ count }} <button @click="count++">+</button>
+      <button @click="count--">-</button> {{ count }} <button @click="count++">+</button>
     </div>
     <div class="count-str">
       <pre class="hug"><code>countStr</code></pre>
@@ -46,7 +46,10 @@ export default defineComponent({
 <style scoped>
   .sync-ref {
     display: flex;
+    flex-wrap: wrap;
     justify-content: space-between;
+    column-gap: 8px;
+    row-gap: 8px;
   }
 
   .hug {
